refactor: move IJwtReplayCache to Features.ReplayPrevention with deprecation shim
The replay-cache primitive is what RFC 7523 §5.2 (JWT-bearer assertion)
and RFC 9449 §11.1.5 (DPoP proof) both want. Moving it out of
Features.JwtBearer into Features.ReplayPrevention removes the
cross-feature coupling new consumers would inherit.

Backward compat via shims at the old location: deprecated interface
derives from canonical, deprecated class derives from canonical impl,
both [Obsolete]. Each class registers as its own DI singleton;
replay-state stays consistent because both delegate to the same
IDistributedCache singleton with a stable cache-key prefix.

Ref #100 #103
