+++ Abblix.Oidc.Server/Features/ReplayPrevention/IJwtReplayCache.cs
@@ -1,0 +1,57 @@
+// Abblix OIDC Server Library
+// Copyright (c) Abblix LLP. All rights reserved.
+//
+// DISCLAIMER: This software is provided 'as-is', without any express or implied
+// warranty. Use at your own risk. Abblix LLP is not liable for any damages
+// arising from the use of this software.
+//
+// LICENSE RESTRICTIONS: This code may not be modified, copied, or redistributed
+// in any form outside of the official GitHub repository at:
+// https://github.com/Abblix/OIDC.Server. All development and modifications
+// must occur within the official repository and are managed solely by Abblix LLP.
+//
+// Unauthorized use, modification, or distribution of this software is strictly
+// prohibited and may be subject to legal action.
+//
+// For full licensing terms, please visit:
+//
+// https://oidc.abblix.com/license
+//
+// CONTACT: For license inquiries or permissions, contact Abblix LLP at
+// info@abblix.com
+
+namespace Abblix.Oidc.Server.Features.ReplayPrevention;
+
+/// <summary>
+/// Tracks JWT IDs (jti claims) presented to the server, so a JWT-bearing flow can detect
+/// replay attempts. Both RFC 7523 §5.2 (JWT-bearer-grant assertion replay) and RFC 9449
+/// §11.1.5 (DPoP proof replay) want this primitive; it is intentionally namespace-neutral
+/// so a single distributed-cache instance serves every consumer.
+/// </summary>
+/// <remarks>
+/// Implementations should use distributed storage (e.g., Redis) so multi-instance
+/// deployments share the replay-protection state.
+/// </remarks>
+public interface IJwtReplayCache
+{
+    /// <summary>
+    /// Checks if a JWT with the specified JTI has already been used.
+    /// </summary>
+    /// <param name="jti">The JWT ID (jti claim) to check.</param>
+    /// <returns>
+    /// A task that completes with true if the JWT has already been used (replay detected);
+    /// false if this is the first time the JWT is being presented.
+    /// </returns>
+    Task<bool> IsReplayedAsync(string jti);
+
+    /// <summary>
+    /// Marks a JWT as used by storing its JTI in the cache until the specified expiration time.
+    /// </summary>
+    /// <param name="jti">The JWT ID (jti claim) to mark as used.</param>
+    /// <param name="expiresAt">
+    /// The time at which the JWT expires. The JTI will be stored until this time plus a small buffer.
+    /// If null, a default expiration will be used.
+    /// </param>
+    /// <returns>A task that completes when the JTI has been stored.</returns>
+    Task MarkAsUsedAsync(string jti, DateTimeOffset? expiresAt);
+}
